Move CRUD module parsers into modulelistparsers
