+++ src/main/java/seedu/address/logic/parser/modulelistparsers/ViewModuleParser.java
@@ -1,9 +1,13 @@
-package seedu.address.logic.parser;
+package seedu.address.logic.parser.modulelistparsers;
 
 import static seedu.address.commons.core.Messages.MESSAGE_INVALID_COMMAND_FORMAT;
 import static seedu.address.logic.parser.CliSyntax.PREFIX_NAME;
 
 import seedu.address.logic.commands.modulelistcommands.ViewModuleCommand;
+import seedu.address.logic.parser.ArgumentMultimap;
+import seedu.address.logic.parser.ArgumentTokenizer;
+import seedu.address.logic.parser.Parser;
+import seedu.address.logic.parser.ParserUtil;
 import seedu.address.logic.parser.exceptions.ParseException;
 import seedu.address.model.module.ModuleName;
