+++ src/test/java/seedu/address/logic/parser/EditModuleParserTest.java
@@ -21,6 +21,7 @@ import org.junit.jupiter.api.Test;
 
 import seedu.address.commons.core.index.Index;
 import seedu.address.logic.commands.modulelistcommands.EditModuleCommand;
+import seedu.address.logic.parser.modulelistparsers.EditModuleParser;
 //import seedu.address.model.person.Email;
 //import seedu.address.model.person.Name;
 //import seedu.address.model.tag.Tag;
